backend/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-CORS==4.0.0
Flask-Marshmallow==1.2.1
marshmallow==3.21.3
marshmallow-sqlalchemy==1.0.0
psycopg[binary]==3.2.13
python-dotenv==1.0.1
